apexcoach-backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
mcap>=1.1.0
mcap-ros2-support>=0.4.0
numpy>=1.26.0
scipy>=1.12.0
shapely>=2.0.0
httpx>=0.27.0
openai>=1.14.0
python-dotenv>=1.0.0
